examples/vocode/pyproject.toml
[project]
name = "voiceobs-vocode-example"
version = "0.1.0"
description = "Example: Vocode Streaming Conversation with voiceobs instrumentation"
requires-python = ">=3.10,<4.0"
dependencies = [
    "voiceobs",
    "vocode>=0.1.0a0",
    "pydantic-settings>=2.0.0",
    "deepgram-sdk>=3.0.0",
]

[tool.uv]
prerelease = "allow"

[dependency-groups]
dev = []

[tool.uv.sources]
voiceobs = { path = "../../" }
